# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Network School OAuth Provider",
    "description": "OAuth 2.0 / OpenID Connect identity provider for Network School.\n\nLets third-party apps request NS user data via standard OAuth scopes \u2014 the same pattern as \"Sign in with Google.\"\n\n## Flows\n\n- **Authorization Code + PKCE** \u2014 user-facing \"Sign in with Network School\" flow\n- **Client Credentials** \u2014 machine-to-machine, no user context\n\n## Scopes\n\n| Scope | Claims |\n|-------|--------|\n| `openid` | `sub` |\n| `email` | `email`, `email_verified` |\n| `profile` | `name`, `picture` |\n| `roles` | `roles` (Discord roles, live) |\n| `date_joined` | `date_joined` |\n| `offline_access` | refresh tokens |\n\n## Authentication\n\n- **User auth**: Discord OAuth2 (NS guild membership required)\n- **Sessions**: HS256 JWT in httponly cookie (`ns_session`)\n- **Access tokens**: RS256 JWT (auto-generated RSA keys)\n- **Client auth**: `client_secret` verified via bcrypt\n",
    "license": {
      "name": "Private",
      "url": "https://ns.com/"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/apps/": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "List all OAuth apps",
        "description": "Returns all registered OAuth applications, sorted by creation date (newest first). Does not include client secrets.",
        "operationId": "list_apps_api_apps__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OAuthAppResponse"
                  },
                  "type": "array",
                  "title": "Response List Apps Api Apps  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "apps"
        ],
        "summary": "Register a new OAuth app",
        "description": "Creates a new OAuth client application. Returns the `client_id` and `client_secret` \u2014 the secret is only returned once at creation time (stored as a bcrypt hash). Scopes are validated against the set of available scopes.",
        "operationId": "create_app_api_apps__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthAppCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "App created. **`client_secret` is only returned in this response.**",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAppCreated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid scopes."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/apps/{app_id}/icon": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Upload app icon",
        "description": "Upload an icon for an OAuth app. Accepts PNG, JPEG, GIF, and WebP. Max file size: 2 MB. Replaces the existing icon if one exists.",
        "operationId": "upload_icon_api_apps__app_id__icon_post",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_icon_api_apps__app_id__icon_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated app with `icon_url`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAppResponse"
                }
              }
            }
          },
          "400": {
            "description": "File type not allowed or file too large."
          },
          "404": {
            "description": "App not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "uploads"
        ],
        "summary": "Delete app icon",
        "description": "Removes the icon from an OAuth app and deletes the file from storage.",
        "operationId": "delete_icon_api_apps__app_id__icon_delete",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "App Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated app with `icon_url: null`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthAppResponse"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_icon_api_apps__app_id__icon_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_icon_api_apps__app_id__icon_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OAuthAppCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name of the application.",
            "examples": [
              "My Cool App"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Short description of the application.",
            "examples": [
              "A demo OAuth client for testing."
            ]
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes",
            "description": "Requested OAuth scopes. Validated against available scopes: openid, profile, email, roles, date_joined, offline_access.",
            "default": [],
            "examples": [
              [
                "openid",
                "email",
                "profile"
              ]
            ]
          },
          "redirect_uris": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Redirect Uris",
            "description": "Allowed OAuth callback URLs.",
            "default": [],
            "examples": [
              [
                "https://myapp.example.com/callback"
              ]
            ]
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url",
            "description": "URL to the app icon (set via icon upload endpoint)."
          },
          "privacy_policy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Policy Url",
            "description": "URL to the app's privacy policy.",
            "examples": [
              "https://myapp.example.com/privacy"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OAuthAppCreate",
        "description": "Request body for creating a new OAuth application."
      },
      "OAuthAppCreated": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique app identifier."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "App description."
          },
          "client_id": {
            "type": "string",
            "title": "Client Id",
            "description": "OAuth client ID (32-char hex)."
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes",
            "description": "Registered scopes."
          },
          "redirect_uris": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Redirect Uris",
            "description": "Registered callback URLs."
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url",
            "description": "App icon URL path."
          },
          "privacy_policy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Policy Url",
            "description": "Privacy policy URL."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Approval status: pending, approved, or rejected.",
            "default": "pending"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "UTC creation timestamp."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "UTC last-updated timestamp."
          },
          "client_secret": {
            "type": "string",
            "title": "Client Secret",
            "description": "Client secret (returned ONCE at creation \u2014 stored as bcrypt hash, never retrievable again)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "client_id",
          "scopes",
          "redirect_uris",
          "created_at",
          "updated_at",
          "client_secret"
        ],
        "title": "OAuthAppCreated",
        "description": "OAuth application with client secret. The secret is only returned once at creation time."
      },
      "OAuthAppResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique app identifier."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "App description."
          },
          "client_id": {
            "type": "string",
            "title": "Client Id",
            "description": "OAuth client ID (32-char hex)."
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes",
            "description": "Registered scopes."
          },
          "redirect_uris": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Redirect Uris",
            "description": "Registered callback URLs."
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url",
            "description": "App icon URL path."
          },
          "privacy_policy_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Policy Url",
            "description": "Privacy policy URL."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Approval status: pending, approved, or rejected.",
            "default": "pending"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "UTC creation timestamp."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "UTC last-updated timestamp."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "client_id",
          "scopes",
          "redirect_uris",
          "created_at",
          "updated_at"
        ],
        "title": "OAuthAppResponse",
        "description": "OAuth application details (without client secret)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}